feat: 유저 정보 수정 엔드포인트 변경
